requirements.txt
tensorflow
opencv-python
numpy
matplotlib
streamlit
scikit-learn
pillow